Fix diff opts handling (ignore case / whitespace)
Inside `vdiff-refresh`, when append is used to build up the diff
command, if one of the intermediate args to append is a string and not
a list, append will expand it into the string's character values.

So, "-w" becomes 45 119, and this `Wrong type argument: stringp, 45`
error message is shown.

Added a small helper that converts non-empty strings using
`ensure-list` to avoid this problem. We only want non-empty strings,
though, otherwise an empty diff opt would get added as an explicit
empty string arg which would cause diff to fail, e.g.

    diff -w "" -- file1 file2

Where we actually want:

    diff -w -- file1 file2

Fixes: #36